🏗️ Initial App Router architecture upgrades, homepage only loading
