+++ README.md
@@ -20,7 +20,6 @@ This is a react app to talk a lil' about me, and for me to add some project exam
   - [Bootstrap](https://getbootstrap.com/docs/5.3/getting-started/introduction/)
 - 📦 Packages
   -  [Sass](https://sass-lang.com/documentation/)
-  - [`react-loader-spinner`](https://mhnpd.github.io/react-loader-spinner/)
   - [`react-device-detect`](https://www.npmjs.com/package/react-device-detect)
 - 🍱 Assets
   - [Devicon Icons](https://devicon.dev/)
