+++ app/layout.tsx
@@ -1,21 +1,20 @@
-import React from 'react'
-import { Html, Head, Main, NextScript } from 'next/document';
+import '@/styles/global.scss'
 
-// any <head> code can go here
-const Document = () => {
+export default function RootLayout({
+    // Layouts must accept a children prop.
+    // This will be populated with nested layouts or pages
+    children,
+}: {
+    children: React.ReactNode
+}) {
     return (
-        <Html lang="en">
-            <Head>
+        <html lang="en">
+            <head>
                 <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0-beta1/dist/css/bootstrap.min.css"
                     integrity="sha384-0evHe/X+R7YkIZDRvuzKMRqM+OrBnVFBL6DOitfPri4tjfHxaWutUpFmBp4vmVor" crossOrigin="anonymous" />
                 <link rel="stylesheet" href="https://cdn.jsdelivr.net/gh/devicons/devicon@v2.15.1/devicon.min.css"></link>
-            </Head>
-            <body>
-                <Main />
-                <NextScript />
-            </body>
-        </Html>
+            </head>
+            <body>{children}</body>
+        </html>
     )
 }
-
-export default Document;
