+++ app/metadata.ts
@@ -1,0 +1,6 @@
+import type { Metadata } from 'next'
+
+export const metadata: Metadata = {
+    title: 'J.L. White',
+    description: 'Jeslyn L. White\'s personal developer website, in case you wanted to learn more about her.'
+}
