+++ components/homeHeader/homeHeader.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/styles.scss';
+@import '../../styles/styles.scss';
 
 .header {
     @include fade-in;
