+++ components/myFooter/myFooter.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/styles.scss';
+@import '../../styles/styles.scss';
 
 .myFooter {
   @include body-text;
