+++ components/myLinks/myLinks.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/styles.scss';
+@import '../../styles/styles.scss';
 
 .linksSection {
     height: 65vh;
