+++ next.config.js
@@ -1,3 +1,5 @@
 module.exports = {
-
+    experimental: {
+        appDir: true,
+    },
 }
